Added TempConverter features
Added code that will catch exceptions such as the user inputting strings
instead of integers.

When an exception happens, it will return to the start of the program.

Added else statement for number not being part of the choices, this will
return the user to the main menu as well.

Main menu is put inside a function instead of being part of the main
method.

When the user inputs a string instead of a number in the part where a
temperature value is asked, it will return them to the menu.